requirements.txt
python-dotenv
langchain
langchain-community
langchain-experimental
openai
pinecone-client
streamlit
streamlit-chat
langchain-pinecone
langchain-openai
assemblyai
mysql-connector-python
pathlib
pymysql
pypdf2
openpyxl
